requirements.txt
python-dotenv
requests
pytz
google-generativeai
Pillow>=10.0.0
